Parametrise the connect functions with the logger to use.
Previously logging was unavoidable, and was to stdout. Now there are
variants of the 'connect' and 'connectWithTLS' functions which take a
logger as an argument, and the default behaviour is to not log at all.

There are now three loggers:

 - stdoutLogger, writes to stdout the same as before
 - fileLogger, appends to a file in the same format as stdoutLogger
 - noopLogger (the new default), does nothing.

In addition to making logging optional, this makes it possible for the
user to write their own loggers.

Closes #2